+++ CHANGELOG.md
@@ -1,8 +1,8 @@
-## [v0.1.1] - 2023-xx-xx
+## [v0.1.1] - 2023-07-12
 - Bug fix: DescriptorEditor image_info address error caused by adding elements to std::vector
 - Bug fix: Pan mode movement
 - Add unlit material
-- Add PBR material local illumination
+- Add PBR material + IBL
 
 ## [v0.1.0] - 2023-07-09
 - Refactor render system
@@ -12,7 +12,6 @@
 - Implement user input
 
 ## [v0.0.1] - 2023-06-18
-
 - Base Vulkan pipeline
 - Monkey head example
 - Console variable system